+++ src/main/java/com/ll/exam/sbb/Question/QuestionService.java
@@ -26,7 +26,7 @@ public class QuestionService {
     return questionRepository.findAll(pageable);
   }
 
-  public Question getQuestion(int id) {
+  public Question getQuestion(long id) {
     return questionRepository.findById(id).orElseThrow(() -> new DataNotFoundException("no %d question not found".formatted(id)));
   }
